% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_viewer_utils.r
\name{add_group_modal_dependency}
\alias{add_group_modal_dependency}
\title{Add Group Modal Dependency}
\usage{
add_group_modal_dependency()
}
\value{
An htmltools htmlDependency object.
}
\description{
Returns the HTML dependency for the add group modal JS and CSS assets.
}
\keyword{internal}
