% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_viewer_utils.r
\name{add_strategy_modal_dependency}
\alias{add_strategy_modal_dependency}
\title{Add Strategy Modal Dependency}
\usage{
add_strategy_modal_dependency()
}
\value{
An htmltools htmlDependency object.
}
\description{
Returns the HTML dependency for the add strategy modal JS and CSS assets.
}
\keyword{internal}
