% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_viewer_utils.r
\name{apply_dsa_params}
\alias{apply_dsa_params}
\title{Apply DSA Parameters to a Model}
\usage{
apply_dsa_params(model, params)
}
\arguments{
\item{model}{An openqaly model object.}

\item{params}{A list of DSA parameter lists (as returned by
\code{normalize_dsa_params}).}
}
\value{
The model with DSA parameters added.
}
\description{
Takes a list of DSA parameter specifications and adds them to a model
using \code{openqaly::add_dsa_variable} or \code{openqaly::add_dsa_setting}.
}
\keyword{internal}
