% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_viewer_results.r
\name{costsResultsServer}
\alias{costsResultsServer}
\title{Costs Results Server}
\usage{
costsResultsServer(id, results, metadata)
}
\arguments{
\item{id}{Module namespace ID.}

\item{results}{Reactive containing model results.}

\item{metadata}{Reactive containing model metadata.}
}
\description{
Costs Results Server
}
\keyword{internal}
