% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_viewer_results.r
\name{costsResultsSidebarUI}
\alias{costsResultsSidebarUI}
\title{Costs Results UI}
\usage{
costsResultsSidebarUI(id)
}
\arguments{
\item{id}{Module namespace ID.}
}
\description{
Costs Results UI
}
\keyword{internal}
