% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_viewer_diagnostics.r
\name{decisionTreeResultsServer}
\alias{decisionTreeResultsServer}
\title{Decision Tree Results Server}
\usage{
decisionTreeResultsServer(id, results, metadata)
}
\arguments{
\item{id}{Module namespace ID.}

\item{results}{Reactive containing model results.}

\item{metadata}{Reactive containing model metadata.}
}
\description{
Decision Tree Results Server
}
\keyword{internal}
