% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_viewer_diagnostics.r
\name{decisionTreeResultsSidebarUI}
\alias{decisionTreeResultsSidebarUI}
\title{Decision Tree Results UI}
\usage{
decisionTreeResultsSidebarUI(id)
}
\arguments{
\item{id}{Module namespace ID.}
}
\description{
Decision Tree Results UI
}
\keyword{internal}
