% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_viewer_psa.r
\name{decode_psa_parameter_choices}
\alias{decode_psa_parameter_choices}
\title{Decode PSA parameter tuple selections from the UI}
\usage{
decode_psa_parameter_choices(selections)
}
\arguments{
\item{selections}{Character vector of encoded tuple selections.}
}
\description{
Decode PSA parameter tuple selections from the UI
}
\keyword{internal}
