% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_viewer_diff.r
\name{diffResultsServer}
\alias{diffResultsServer}
\title{Diff Results Server}
\usage{
diffResultsServer(id, model, current_model)
}
\arguments{
\item{id}{Module namespace ID.}

\item{model}{Reactive containing the original model.}

\item{current_model}{Reactive containing the current model with all UI edits applied.}
}
\description{
Diff Results Server
}
\keyword{internal}
