% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_viewer_utils.r
\name{documentation_editor_dependency}
\alias{documentation_editor_dependency}
\title{Documentation Editor Dependency}
\usage{
documentation_editor_dependency()
}
\value{
An htmltools htmlDependency object.
}
\description{
Returns the HTML dependency for the documentation editor JS and CSS assets.
}
\keyword{internal}
