% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_suggestions.r
\name{.build_keyword_suggestions}
\alias{.build_keyword_suggestions}
\title{Build keyword suggestions data frame}
\usage{
.build_keyword_suggestions(keywords)
}
\description{
Build keyword suggestions data frame
}
\keyword{internal}
