% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_suggestions.r
\name{.build_signature}
\alias{.build_signature}
\title{Build function signature from formals}
\usage{
.build_signature(fn_name, fn)
}
\arguments{
\item{fn_name}{Function name}

\item{fn}{Function object}
}
\value{
Character string of function signature
}
\description{
Build function signature from formals
}
\keyword{internal}
