% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_suggestions.r
\name{.build_variable_suggestions}
\alias{.build_variable_suggestions}
\title{Build variable suggestions with strategy/group-specific display}
\usage{
.build_variable_suggestions(model)
}
\description{
Build variable suggestions with strategy/group-specific display
}
\keyword{internal}
