% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_suggestions.r
\name{.empty_suggestions_df}
\alias{.empty_suggestions_df}
\title{Create empty suggestions data frame}
\usage{
.empty_suggestions_df()
}
\description{
Create empty suggestions data frame
}
\keyword{internal}
