% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_suggestions.r
\name{.extract_help_description}
\alias{.extract_help_description}
\title{Extract description from help file}
\usage{
.extract_help_description(fn_name, pkg)
}
\arguments{
\item{fn_name}{Function name}

\item{pkg}{Package name}
}
\value{
Character string of function description or NA_character_
}
\description{
Extract description from help file
}
\keyword{internal}
