% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_suggestions.r
\name{.extract_variable_names}
\alias{.extract_variable_names}
\title{Extract variable names from model}
\usage{
.extract_variable_names(model)
}
\description{
Extract variable names from model
}
\keyword{internal}
