% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/override_input.r
\name{.parse_timeframe_value}
\alias{.parse_timeframe_value}
\title{Parse Timeframe Value}
\usage{
.parse_timeframe_value(value)
}
\arguments{
\item{value}{A string in "number|unit" format.}
}
\value{
A list with \code{number} and \code{unit} elements.
}
\description{
Splits a timeframe value string in "number|unit" format.
}
\keyword{internal}
