% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_viewer_utils.r
\name{dsa_params_dependency}
\alias{dsa_params_dependency}
\title{DSA Params Table Dependency}
\usage{
dsa_params_dependency()
}
\value{
An htmltools htmlDependency object.
}
\description{
Returns the HTML dependency for the DSA parameter table JS and CSS assets.
}
\keyword{internal}
