% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_viewer_dsa.r
\name{dsaResultsServer}
\alias{dsaResultsServer}
\title{DSA Results Server}
\usage{
dsaResultsServer(id, dsa_results, metadata)
}
\arguments{
\item{id}{Module namespace ID.}

\item{dsa_results}{Reactive containing DSA results.}

\item{metadata}{Reactive containing model metadata.}
}
\description{
DSA Results Server
}
\keyword{internal}
