% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_viewer_dsa.r
\name{dsaResultTabServer}
\alias{dsaResultTabServer}
\title{DSA Result Tab Server}
\usage{
dsaResultTabServer(id, analysis_type, dsa_results, metadata)
}
\arguments{
\item{id}{Module namespace ID.}

\item{analysis_type}{Fixed string: "outcomes", "costs", "nmb", "ce", or "vbp".}

\item{dsa_results}{Reactive containing DSA results.}

\item{metadata}{Reactive containing model metadata.}
}
\description{
DSA Result Tab Server
}
\keyword{internal}
