% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_viewer_dsa.r
\name{dsaResultTabSidebarUI}
\alias{dsaResultTabSidebarUI}
\title{DSA Result Tab UI}
\usage{
dsaResultTabSidebarUI(id)
}
\arguments{
\item{id}{Module namespace ID.}
}
\description{
DSA Result Tab UI
}
\keyword{internal}
