% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_viewer_psa.r
\name{encode_psa_parameter_choice}
\alias{encode_psa_parameter_choice}
\title{Encode a PSA parameter tuple as a select input value}
\usage{
encode_psa_parameter_choice(variable, strategy = NULL, group = NULL)
}
\arguments{
\item{variable}{Variable name.}

\item{strategy}{Strategy name.}

\item{group}{Group name.}
}
\description{
Encode a PSA parameter tuple as a select input value
}
\keyword{internal}
