% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_viewer_utils.r
\name{get_cost_summary_choices}
\alias{get_cost_summary_choices}
\title{Get Cost Summary Choices}
\usage{
get_cost_summary_choices(metadata)
}
\arguments{
\item{metadata}{A metadata list from openqaly model results.}
}
\value{
A named character vector of cost summary choices.
}
\description{
Extract available cost summaries from model results metadata.
}
\keyword{internal}
