% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_viewer_utils.r
\name{get_group_choices}
\alias{get_group_choices}
\title{Get Group Choices}
\usage{
get_group_choices(metadata)
}
\arguments{
\item{metadata}{A metadata list from openqaly model results.}
}
\value{
A character vector of group names.
}
\description{
Extract available group names from model results metadata.
}
\keyword{internal}
