% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_viewer_utils.r
\name{get_outcome_summary_choices}
\alias{get_outcome_summary_choices}
\title{Get Outcome Summary Choices}
\usage{
get_outcome_summary_choices(metadata)
}
\arguments{
\item{metadata}{A metadata list from openqaly model results.}
}
\value{
A named character vector of outcome summary choices.
}
\description{
Extract available outcome summaries from model results metadata.
}
\keyword{internal}
