% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_viewer_utils.r
\name{groups_table_dependency}
\alias{groups_table_dependency}
\title{Groups Table Dependency}
\usage{
groups_table_dependency()
}
\value{
An htmltools htmlDependency object.
}
\description{
Returns the HTML dependency for the groups table JS and CSS assets.
}
\keyword{internal}
