% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_viewer_results.r
\name{incrementalCeResultsServer}
\alias{incrementalCeResultsServer}
\title{Incremental CE Results Server}
\usage{
incrementalCeResultsServer(id, results, metadata)
}
\arguments{
\item{id}{Module namespace ID.}

\item{results}{Reactive containing model results.}

\item{metadata}{Reactive containing model metadata.}
}
\description{
Incremental CE Results Server
}
\keyword{internal}
