% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_viewer_results.r
\name{incrementalCeResultsSidebarUI}
\alias{incrementalCeResultsSidebarUI}
\title{Incremental CE Results UI}
\usage{
incrementalCeResultsSidebarUI(id)
}
\arguments{
\item{id}{Module namespace ID.}
}
\description{
Incremental CE Results UI
}
\keyword{internal}
