% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_viewer_results.r
\name{incrementalCeResultsUI}
\alias{incrementalCeResultsUI}
\title{Incremental CE Results UI}
\usage{
incrementalCeResultsUI(id)
}
\arguments{
\item{id}{Module namespace ID.}
}
\description{
Incremental CE Results UI
}
\keyword{internal}
