% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_viewer_utils.r
\name{info_hover_dependency}
\alias{info_hover_dependency}
\title{Info Hover Dependency}
\usage{
info_hover_dependency()
}
\value{
An htmltools htmlDependency object.
}
\description{
Returns the HTML dependency for the info hover popover CSS and JS.
}
\keyword{internal}
