% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_viewer_psa.r
\name{map_psa_group_name}
\alias{map_psa_group_name}
\title{Map a group name to its display name}
\usage{
map_psa_group_name(group_name, metadata)
}
\description{
Map a group name to its display name
}
\keyword{internal}
