% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_viewer_psa.r
\name{map_psa_strategy_name}
\alias{map_psa_strategy_name}
\title{Map a strategy name to its display name}
\usage{
map_psa_strategy_name(strategy_name, metadata)
}
\description{
Map a strategy name to its display name
}
\keyword{internal}
