% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_suggestions.r
\name{model_suggestions}
\alias{model_suggestions}
\title{Model-Aware Formula Suggestions}
\description{
Functions for extracting context-appropriate terms and suggestions from
openqaly models for use with \code{\link{formulaInput}}.
}
