% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_viewer_results.r
\name{nmbResultsServer}
\alias{nmbResultsServer}
\title{NMB Results Server}
\usage{
nmbResultsServer(id, results, metadata)
}
\arguments{
\item{id}{Module namespace ID.}

\item{results}{Reactive containing model results.}

\item{metadata}{Reactive containing model metadata.}
}
\description{
NMB Results Server
}
\keyword{internal}
