% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_viewer_results.r
\name{nmbResultsSidebarUI}
\alias{nmbResultsSidebarUI}
\title{NMB Results UI}
\usage{
nmbResultsSidebarUI(id)
}
\arguments{
\item{id}{Module namespace ID.}
}
\description{
NMB Results UI
}
\keyword{internal}
