% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_viewer_results.r
\name{outcomesResultsServer}
\alias{outcomesResultsServer}
\title{Outcomes Results Server}
\usage{
outcomesResultsServer(id, results, metadata)
}
\arguments{
\item{id}{Module namespace ID.}

\item{results}{Reactive containing model results.}

\item{metadata}{Reactive containing model metadata.}
}
\description{
Outcomes Results Server
}
\keyword{internal}
