% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_viewer_results.r
\name{outcomesResultsSidebarUI}
\alias{outcomesResultsSidebarUI}
\title{Outcomes Results UI}
\usage{
outcomesResultsSidebarUI(id)
}
\arguments{
\item{id}{Module namespace ID.}
}
\description{
Outcomes Results UI
}
\keyword{internal}
