% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/override_input.r
\name{override_input_dependency}
\alias{override_input_dependency}
\title{Override Input Dependency}
\usage{
override_input_dependency()
}
\value{
An htmltools htmlDependency object.
}
\description{
Returns the HTML dependency for override input JS and CSS assets.
}
\keyword{internal}
