% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/override_manager.r
\name{override_manager_dependency}
\alias{override_manager_dependency}
\title{Override Manager Dependency}
\usage{
override_manager_dependency()
}
\value{
A list of htmltools htmlDependency objects.
}
\description{
Returns the HTML dependencies for the override manager, including SortableJS
from CDN and local manager CSS/JS assets.
}
\keyword{internal}
