% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/override_input.r
\name{overrideInput}
\alias{overrideInput}
\title{Override Input}
\usage{
overrideInput(inputId, model, width = NULL)
}
\arguments{
\item{inputId}{The base input ID. Individual overrides will be accessible
as \code{input$\{inputId\}_{override_name}}.}

\item{model}{An openqaly model object that contains \code{override_categories}.}

\item{width}{The width of the container, e.g., '400px' or '100\\%'.
Default is NULL which uses the default width.}
}
\value{
A Shiny tag list containing a tabbed interface with override cards.
}
\description{
Generates a tabbed interface for model override categories, with each
override displayed as a card containing the appropriate input control.
This is a container generator - each override becomes its own independent
Shiny input accessible via \code{input$\{inputId\}_{name}}.
}
\examples{
\dontrun{
# Run the interactive demo:
run_override_input_demo()
}

}
\seealso{
\code{\link{updateOverrideInput}}, \code{\link{run_override_input_demo}}
}
