% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_viewer_results.r
\name{pairwiseCeResultsServer}
\alias{pairwiseCeResultsServer}
\title{Pairwise CE Results Server}
\usage{
pairwiseCeResultsServer(id, results, metadata)
}
\arguments{
\item{id}{Module namespace ID.}

\item{results}{Reactive containing model results.}

\item{metadata}{Reactive containing model metadata.}
}
\description{
Pairwise CE Results Server
}
\keyword{internal}
