% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_viewer_results.r
\name{pairwiseCeResultsUI}
\alias{pairwiseCeResultsUI}
\title{Pairwise CE Results UI}
\usage{
pairwiseCeResultsUI(id)
}
\arguments{
\item{id}{Module namespace ID.}
}
\description{
Pairwise CE Results UI
}
\keyword{internal}
