% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_viewer_utils.r
\name{psa_params_dependency}
\alias{psa_params_dependency}
\title{PSA Parameters Dependency}
\usage{
psa_params_dependency()
}
\value{
An htmltools htmlDependency object.
}
\description{
Returns the HTML dependency for the PSA parameter table JS and CSS assets.
}
\keyword{internal}
