% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_viewer_psa.r
\name{psaResultTabServer}
\alias{psaResultTabServer}
\title{PSA Result Tab Server}
\usage{
psaResultTabServer(id, analysis_type, psa_results, metadata)
}
\arguments{
\item{id}{Module namespace ID.}

\item{analysis_type}{Fixed string: "outcomes", "costs", "nmb",
"incremental_ce", "pairwise_ce", "evpi", or "parameters".}

\item{psa_results}{Reactive containing PSA results.}

\item{metadata}{Reactive containing model metadata.}
}
\description{
PSA Result Tab Server
}
\keyword{internal}
