% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_viewer_psa.r
\name{psaResultTabSidebarUI}
\alias{psaResultTabSidebarUI}
\title{PSA Result Tab UI}
\usage{
psaResultTabSidebarUI(id)
}
\arguments{
\item{id}{Module namespace ID.}
}
\description{
PSA Result Tab UI
}
\keyword{internal}
