% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_viewer_utils.r
\name{render_flextable_html}
\alias{render_flextable_html}
\title{Render Flextable as HTML}
\usage{
render_flextable_html(ft)
}
\arguments{
\item{ft}{A flextable object.}
}
\value{
An htmltools tag object.
}
\description{
Converts a flextable object to an HTML widget suitable for Shiny renderUI.
}
\keyword{internal}
