% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_viewer_utils.r
\name{scenario_params_dependency}
\alias{scenario_params_dependency}
\title{Scenario Parameters Dependency}
\usage{
scenario_params_dependency()
}
\value{
An htmltools htmlDependency object.
}
\description{
Returns the HTML dependency for the scenario parameters JS and CSS assets.
}
\keyword{internal}
