% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_viewer_scenario.r
\name{scenarioResultTabServer}
\alias{scenarioResultTabServer}
\title{Scenario Result Tab Server}
\usage{
scenarioResultTabServer(id, analysis_type, scenario_results, metadata)
}
\arguments{
\item{id}{Module namespace ID.}

\item{analysis_type}{Fixed string: "outcomes", "costs", "nmb", "ce", or "vbp".}

\item{scenario_results}{Reactive containing scenario results.}

\item{metadata}{Reactive containing model metadata.}
}
\description{
Scenario Result Tab Server
}
\keyword{internal}
