% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_viewer_scenario.r
\name{scenarioResultTabUI}
\alias{scenarioResultTabUI}
\title{Scenario Result Tab UI}
\usage{
scenarioResultTabUI(id)
}
\arguments{
\item{id}{Module namespace ID.}
}
\description{
Scenario Result Tab UI
}
\keyword{internal}
