% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_viewer_utils.r
\name{states_table_dependency}
\alias{states_table_dependency}
\title{States Table Dependency}
\usage{
states_table_dependency()
}
\value{
An htmltools htmlDependency object.
}
\description{
Returns the HTML dependency for the states table JS and CSS assets.
}
\keyword{internal}
