% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_viewer_utils.r
\name{strategies_table_dependency}
\alias{strategies_table_dependency}
\title{Strategies Table Dependency}
\usage{
strategies_table_dependency()
}
\value{
An htmltools htmlDependency object.
}
\description{
Returns the HTML dependency for the strategies table JS and CSS assets.
}
