% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_viewer_utils.r
\name{summaries_table_dependency}
\alias{summaries_table_dependency}
\title{Summaries Table Dependency}
\usage{
summaries_table_dependency()
}
\value{
An htmltools htmlDependency object.
}
\description{
Returns the HTML dependency for the summaries table JS and CSS assets.
}
\keyword{internal}
