% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_viewer_utils.r
\name{threshold_params_dependency}
\alias{threshold_params_dependency}
\title{Threshold Parameters Dependency}
\usage{
threshold_params_dependency()
}
\value{
An htmltools htmlDependency object.
}
\description{
Returns the HTML dependency for the threshold parameter table JS and CSS assets.
}
