% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_viewer_threshold.r
\name{thresholdResultTabServer}
\alias{thresholdResultTabServer}
\title{Threshold Result Tab Server}
\usage{
thresholdResultTabServer(id, tab_type, threshold_results)
}
\arguments{
\item{id}{Module namespace ID.}

\item{tab_type}{Fixed string: "detail" or "convergence".}

\item{threshold_results}{Reactive containing threshold results.}
}
\description{
Threshold Result Tab Server
}
