% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_viewer_threshold.r
\name{thresholdResultTabUI}
\alias{thresholdResultTabUI}
\title{Threshold Result Tab UI}
\usage{
thresholdResultTabUI(id)
}
\arguments{
\item{id}{Module namespace ID.}
}
\description{
Threshold Result Tab UI
}
