% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_viewer_threshold.r
\name{thresholdSummaryServer}
\alias{thresholdSummaryServer}
\title{Threshold Summary Server}
\usage{
thresholdSummaryServer(id, threshold_results)
}
\arguments{
\item{id}{Module namespace ID.}

\item{threshold_results}{Reactive containing threshold results.}
}
\description{
Threshold Summary Server
}
