% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_viewer_threshold.r
\name{thresholdSummaryUI}
\alias{thresholdSummaryUI}
\title{Threshold Summary UI}
\usage{
thresholdSummaryUI(id)
}
\arguments{
\item{id}{Module namespace ID.}
}
\description{
Threshold Summary UI
}
