% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_viewer_results.r
\name{traceResultsServer}
\alias{traceResultsServer}
\title{Trace Results Server}
\usage{
traceResultsServer(id, results, metadata)
}
\arguments{
\item{id}{Module namespace ID.}

\item{results}{Reactive containing model results.}

\item{metadata}{Reactive containing model metadata.}
}
\description{
Trace Results Server
}
\keyword{internal}
