% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_viewer_results.r
\name{traceResultsSidebarUI}
\alias{traceResultsSidebarUI}
\title{Trace Results UI}
\usage{
traceResultsSidebarUI(id)
}
\arguments{
\item{id}{Module namespace ID.}
}
\description{
Trace Results UI
}
\keyword{internal}
