% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_viewer_diagnostics.r
\name{transitionHeatmapSidebarUI}
\alias{transitionHeatmapSidebarUI}
\title{Transition Heatmap UI}
\usage{
transitionHeatmapSidebarUI(id)
}
\arguments{
\item{id}{Module namespace ID.}
}
\description{
Transition Heatmap UI
}
\keyword{internal}
