% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_viewer_utils.r
\name{transitions_table_dependency}
\alias{transitions_table_dependency}
\title{Transitions Table Dependency}
\usage{
transitions_table_dependency()
}
\value{
An htmltools htmlDependency object.
}
\description{
Returns the HTML dependency for the transitions table JS and CSS assets.
}
\keyword{internal}
