% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_viewer_utils.r
\name{trees_table_dependency}
\alias{trees_table_dependency}
\title{Trees Table Dependency}
\usage{
trees_table_dependency()
}
\value{
An htmltools htmlDependency object.
}
\description{
Returns the HTML dependency for the trees table JS and CSS assets.
}
\keyword{internal}
