% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_viewer_utils.r
\name{twsa_params_dependency}
\alias{twsa_params_dependency}
\title{TWSA Parameters Dependency}
\usage{
twsa_params_dependency()
}
\value{
An htmltools htmlDependency object.
}
\description{
Returns the HTML dependency for the TWSA parameter table JS and CSS assets.
}
\keyword{internal}
