% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_viewer_twsa.r
\name{twsaResultTabServer}
\alias{twsaResultTabServer}
\title{TWSA Result Tab Server}
\usage{
twsaResultTabServer(id, analysis_type, twsa_results, metadata)
}
\arguments{
\item{id}{Module namespace ID.}

\item{analysis_type}{Fixed string: "outcomes", "costs", "nmb", "ce", or "vbp".}

\item{twsa_results}{Reactive containing TWSA results.}

\item{metadata}{Reactive containing model metadata.}
}
\description{
TWSA Result Tab Server
}
\keyword{internal}
