% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_viewer_twsa.r
\name{twsaResultTabSidebarUI}
\alias{twsaResultTabSidebarUI}
\title{TWSA Result Tab UI}
\usage{
twsaResultTabSidebarUI(id)
}
\arguments{
\item{id}{Module namespace ID.}
}
\description{
TWSA Result Tab UI
}
\keyword{internal}
