% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_viewer_twsa.r
\name{twsaResultTabUI}
\alias{twsaResultTabUI}
\title{TWSA Result Tab UI}
\usage{
twsaResultTabUI(id)
}
\arguments{
\item{id}{Module namespace ID.}
}
\description{
TWSA Result Tab UI
}
\keyword{internal}
