% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_input.r
\name{updateFormulaInput}
\alias{updateFormulaInput}
\title{Update Formula Input}
\usage{
updateFormulaInput(
  session,
  inputId,
  value = NULL,
  terms = NULL,
  suggestions = NULL
)
}
\arguments{
\item{session}{The session object passed to the Shiny server function.}

\item{inputId}{The id of the input object.}

\item{value}{The new value for the input.}

\item{terms}{A named list of terms to highlight. See \code{\link{formulaInput}}
for details on the format. Set to an empty list \code{list()} to clear
all term highlighting. Use \code{\link{get_model_terms}} to extract from
an openqaly model.}

\item{suggestions}{A named list of autocomplete suggestions. See
\code{\link{formulaInput}} for details on the format. Set to an empty
list \code{list()} to clear all suggestions. Use
\code{\link{get_model_suggestions}} to extract from an openqaly model.}
}
\description{
Change the value, terms, and/or suggestions of a formula input on the client.
}
\examples{
\dontrun{
# Reactive update when model or context changes
observe({
  updateFormulaInput(session, "formula",
    terms = get_model_terms(model(), context()),
    suggestions = get_model_suggestions(model(), context()))
})
}

}
\seealso{
\code{\link{formulaInput}}, \code{\link{get_model_terms}},
\code{\link{get_model_suggestions}}
}
