% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/override_input.r
\name{updateOverrideInput}
\alias{updateOverrideInput}
\title{Update Override Input}
\usage{
updateOverrideInput(
  session,
  inputId,
  override_name,
  value,
  strategy = "",
  group = ""
)
}
\arguments{
\item{session}{The session object passed to the Shiny server function.}

\item{inputId}{The base input ID used when creating the override input.}

\item{override_name}{The name of the specific override to update.}

\item{value}{The new value for the override.}
}
\description{
Update the value of a specific override within an override input container.
}
\examples{
\dontrun{
# In server function:
updateOverrideInput(session, "overrides", "discount_rate", 0.05)
}

}
\seealso{
\code{\link{overrideInput}}
}
