% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_viewer_variable_diagnostics.r
\name{variableDiagnosticsServer}
\alias{variableDiagnosticsServer}
\title{Variable Diagnostics Server}
\usage{
variableDiagnosticsServer(id, results, metadata)
}
\arguments{
\item{id}{Module namespace ID.}

\item{results}{Reactive containing model results.}

\item{metadata}{Reactive containing model metadata.}
}
\description{
Variable Diagnostics Server
}
\keyword{internal}
