% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_viewer_variable_diagnostics.r
\name{variableDiagnosticsSidebarUI}
\alias{variableDiagnosticsSidebarUI}
\title{Variable Diagnostics UI}
\usage{
variableDiagnosticsSidebarUI(id)
}
\arguments{
\item{id}{Module namespace ID.}
}
\description{
Variable Diagnostics UI
}
\keyword{internal}
