% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_viewer_variable_diagnostics.r
\name{variableDiagnosticsUI}
\alias{variableDiagnosticsUI}
\title{Variable Diagnostics UI}
\usage{
variableDiagnosticsUI(id)
}
\arguments{
\item{id}{Module namespace ID.}
}
\description{
Variable Diagnostics UI
}
\keyword{internal}
