% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_viewer_utils.r
\name{variables_table_dependency}
\alias{variables_table_dependency}
\title{Variables Table Dependency}
\usage{
variables_table_dependency()
}
\value{
An htmltools htmlDependency object.
}
\description{
Returns the HTML dependency for the variables table JS and CSS assets.
}
\keyword{internal}
