% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_viewer_vbp.r
\name{vbpResultsServer}
\alias{vbpResultsServer}
\title{VBP Results Server}
\usage{
vbpResultsServer(id, vbp_results, metadata)
}
\arguments{
\item{id}{Module namespace ID.}

\item{vbp_results}{Reactive containing VBP results.}

\item{metadata}{Reactive containing model metadata.}
}
\description{
VBP Results Server
}
\keyword{internal}
