% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_viewer_vbp.r
\name{vbpResultsSidebarUI}
\alias{vbpResultsSidebarUI}
\title{VBP Results UI}
\usage{
vbpResultsSidebarUI(id)
}
\arguments{
\item{id}{Module namespace ID.}
}
\description{
VBP Results UI
}
\keyword{internal}
