% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_viewer_psa.r
\name{encode_psa_parameter_choice}
\alias{encode_psa_parameter_choice}
\title{Normalize PSA parameter group filters for tuple choices}
\usage{
encode_psa_parameter_choice(variable, strategy = NULL, group = NULL)
}
\arguments{
\item{variable}{Variable name.}

\item{strategy}{Strategy name.}

\item{group}{Group name.}

\item{groups_input}{Selected group filter values from the UI.}

\item{metadata}{Optional PSA metadata.}
}
\description{
Normalize PSA parameter group filters for tuple choices
}
\keyword{Encode}
\keyword{PSA}
\keyword{a}
\keyword{as}
\keyword{input}
\keyword{internal}
\keyword{parameter}
\keyword{select}
\keyword{tuple}
\keyword{value}
